AccuracyTestBackend/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.1
dj-database-url==3.0.1
Django==4.2.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-knox==4.2.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
environ==1.0
gunicorn==21.2.0
packaging==25.0
psycopg==3.1.17
pycparser==2.21
pytz==2023.3.post1
sqlparse==0.4.4
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.9.0
whitenoise==6.11.0
